+++ src/screens/Login/Login.screen.js
@@ -16,6 +16,8 @@ import LinearGradient from 'react-native-linear-gradient';
 import FontAwesome from 'react-native-vector-icons/FontAwesome';
 import Feather from 'react-native-vector-icons/Feather';
 
+import styles from './Login.style';
+
 const Login = ({navigation}) => {
   const [data, setData] = useState({
     email: '',
@@ -101,9 +103,11 @@ const Login = ({navigation}) => {
 
         <View style={styles.button}>
           <TouchableOpacity
-            onPress={() => navigation.navigate('Home')}
-            style={styles.signIn}>
-            <LinearGradient colors={['#08d4c4', '#01ab9d']}>
+            style={{width: '100%'}}
+            onPress={() => navigation.navigate('Home')}>
+            <LinearGradient
+              colors={['#08d4c4', '#01ab9d']}
+              style={[styles.signIn]}>
               <Text style={[styles.textSign, {color: 'white'}]}>Login</Text>
             </LinearGradient>
           </TouchableOpacity>
@@ -123,62 +127,3 @@ const Login = ({navigation}) => {
 };
 
 export default Login;
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#009387',
-  },
-  header: {
-    flex: 1,
-    justifyContent: 'flex-end',
-    paddingHorizontal: 20,
-    paddingBottom: 50,
-  },
-  footer: {
-    flex: 3,
-    backgroundColor: '#fff',
-    borderTopLeftRadius: 30,
-    borderTopRightRadius: 30,
-    paddingHorizontal: 20,
-    paddingVertical: 30,
-  },
-
-  text_header: {
-    color: '#fff',
-    fontWeight: 'bold',
-    fontSize: 30,
-  },
-  text_footer: {
-    color: '#05375a',
-    fontSize: 18,
-  },
-  action: {
-    flexDirection: 'row',
-    marginTop: 10,
-    borderBottomWidth: 1,
-    borderBottomColor: '#f2f2f2',
-    paddingBottom: 5,
-  },
-  textInput: {
-    flex: 1,
-    marginTop: Platform.OS == 'ios' ? 0 : -12,
-    paddingLeft: 10,
-    color: '#05375a',
-  },
-  button: {
-    alignItems: 'center',
-    marginTop: 50,
-  },
-  signIn: {
-    width: '100%',
-    height: 50,
-    justifyContent: 'center',
-    alignItems: 'center',
-    borderRadius: 10,
-  },
-  textSign: {
-    fontSize: 18,
-    fontWeight: 'bold',
-  },
-});
